requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
mangum==0.17.0
supabase==2.3.4
gotrue==1.3.0
python-dotenv==1.0.0
pillow==10.1.0
aiofiles==23.2.1
httpx==0.25.2
PyJWT==2.8.0vicorn[standard]==0.24.0
sqlalchemy[postgresql]==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
mangum==0.17.0
supabase==2.3.4
gotrue==1.3.0
python-dotenv==1.0.0
pillow==10.1.0
aiofiles==23.2.1
httpx==0.25.2
